% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_dashboard.R
\encoding{UTF-8}
\name{build_dashboard}
\alias{build_dashboard}
\title{Create IPCA forecast model dashboard}
\usage{
build_dashboard(
  y = "ipca",
  data = ts_transform(get_data(), y),
  accuracy = train_models(data, y),
  forecast = forecast_ensemble(ts_transform(get_data(), y, seasonal_dummies = FALSE), y),
  ensemble_col = "ensemble",
  title_fanchart = rlang::as_utf8_character("**Fanchart**: Previsão do IPCA"),
  y_lab_fanchart = "\% a.m.",
  caption_lab = rlang::as_utf8_character("**Elaboração:** analisemacro.com.br"),
  title_accuracy = rlang::as_utf8_character("**Acurácia**: performance por horizonte"),
  subtitle_accuracy = rlang::as_utf8_character("Modelos de previsão do IPCA"),
  x_lab_accuracy = "Horizonte (meses)",
  save_at = "docs"
)
}
\arguments{
\item{y}{Column name of the variable of interest}

\item{data}{A tsibble object, similar to the output of \code{ts_transform(get_data(), y)}}

\item{accuracy}{A tibble object, similar and with the same column names as the output of \code{train_models}}

\item{forecast}{A tsibble object, similar to the output of \code{forecast_ensemble}}

\item{ensemble_col}{Column name of ensemble forecasts in \code{forecast}}

\item{title_fanchart}{Title for the forecast fanchart}

\item{y_lab_fanchart}{Fanchart Y-axis title}

\item{caption_lab}{Text for caption}

\item{title_accuracy}{Title for predictive horizon accuracy graph (\code{accuracy} data)}

\item{subtitle_accuracy}{Subtitle for the accuracy graph by predictive horizon (\code{accuracy} data)}

\item{x_lab_accuracy}{X-axis text in accuracy graph by predictive horizon (\code{accuracy} data)}

\item{save_at}{Directory name to save the HTML file, if it does not exist it will be created}
}
\value{
Flexdashboard HTML file named as \code{index.html}.
}
\description{
Create IPCA forecast model dashboard
}
\examples{
td <- tempdir()
build_dashboard(save_at = td)
}
