% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_data.R
\name{create_macro}
\alias{create_macro}
\title{indice Geral
Alimentação
Habitação
Artigos de Residência
Vestuário
Transportes
Saúde e Cuidados Pessoais
Despesas Pessoais
Educação
Comunicação
Combinando}
\usage{
create_macro(...)
}
\description{
indice Geral
Alimentação
Habitação
Artigos de Residência
Vestuário
Transportes
Saúde e Cuidados Pessoais
Despesas Pessoais
Educação
Comunicação
Combinando
}
