% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forecast_ensemble.R
\name{forecast_ensemble}
\alias{forecast_ensemble}
\title{Generate an out-of-sample Ensemble forecast from the CSR and Bagging models}
\usage{
forecast_ensemble(
  data,
  y,
  horizon = 12,
  K = 20,
  k = 15,
  R = 500,
  pre_testing = "group-joint",
  seasonal_dummies = TRUE,
  auto_scenario = TRUE,
  xreg_scenario,
  ...
)
}
\arguments{
\item{data}{A tsibble object}

\item{y}{Column name of the variable of interest}

\item{horizon}{Forecast horizon}

\item{K}{Number of variables to be selected after the pre-testing. If K=ncol(x) the pre-testing is redundant (see \code{HDeconometrics::csr}).}

\item{k}{Number of variables in each subset. Must be smaller than K (see \code{HDeconometrics::csr}).}

\item{R}{Number of bootstrap replucations (see \code{HDeconometrics::bagging}).}

\item{pre_testing}{The type of pre-testing (see \code{HDeconometrics::bagging}).}

\item{seasonal_dummies}{If \code{TRUE}, seasonal dummies are created for model estimation and out-of-sample forecasting.}

\item{auto_scenario}{If \code{TRUE}, the \code{forecast::auto.arima} function is used to generate scenarios for the independent variables in \code{horizon} out-of-sample periods. If \code{FALSE}, the user must supply a \code{xreg_scenario} matrix with the scenarios, containing a number of lines equal to \code{horizon}.}

\item{xreg_scenario}{Scenario matrix for independent variables if \verb{auto scenario} is \code{FALSE}.}

\item{...}{Other arguments passed to \code{HDeconometrics::bagging}}
}
\value{
A tibble
}
\description{
Generate an out-of-sample Ensemble forecast from the CSR and Bagging models
}
\examples{
df_macro <- ts_transform(get_data(), "ipca", seasonal_dummies = FALSE)
df_fcsts <- forecast_ensemble(df_macro, "ipca")
}
