% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/train_models.R
\name{get_cv_rmse_hdecon}
\alias{get_cv_rmse_hdecon}
\title{CSR and Bagging: estimates model with cross-validation and reports accuracy by forecast horizon}
\usage{
get_cv_rmse_hdecon(
  model,
  data,
  y,
  init_window = 150,
  horizon = 12,
  quiet = FALSE,
  ...
)
}
\arguments{
\item{model}{Model to be estimated, possible values are \code{csr} or \code{bagging}, see HDeconometrics functions}

\item{data}{A tsibble}

\item{y}{Column name of the variable of interest (used to report accuracy)}

\item{init_window}{Number of initial observations to be used in the first cross-validation subsample}

\item{horizon}{Forecast horizon}

\item{quiet}{The default (\code{FALSE}) prints log messages to the Console.}

\item{...}{Additional arguments to \code{HDeconometrics::csr} or \code{HDeconometrics::bagging}}
}
\value{
List with 2 tibbles: RMSE per forecast horizon and point forecast.
}
\description{
CSR and Bagging: estimates model with cross-validation and reports accuracy by forecast horizon
}
