% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{report_ndiffs}
\alias{report_ndiffs}
\title{Report number of differences to make time series stationary (vectorized)}
\usage{
report_ndiffs(
  x,
  test = c("kpss", "adf", "pp"),
  term = c("level", "trend"),
  alpha = 0.05,
  na_rm = TRUE
)
}
\arguments{
\item{x}{List-like object with vectors of the series to be tested}

\item{test}{Type of unit root test to use, see forecast::ndiffs}

\item{term}{Specification of the deterministic component in the regression, see forecast::ndiffs}

\item{alpha}{Level of the test, possible values range from 0.01 to 0.1}

\item{na_rm}{Remove NAs from x?}
}
\value{
Tibble with variable name from x and the number of differences found
}
\description{
Report number of differences to make time series stationary (vectorized)
}
\examples{
ts1 <- as.vector(arima.sim(list(order = c(1, 1, 0), ar = 0.7), n = 19))
ts2 <- rnorm(20)
report_ndiffs(data.frame(ts1 = ts1, ts2 = ts2))
}
