% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ts_transform.R
\name{ts_transform}
\alias{ts_transform}
\title{Black box function to prepare time series}
\usage{
ts_transform(
  data,
  y,
  stationary = TRUE,
  lags = 1:4,
  fill_down_na = TRUE,
  seasonal_dummies = TRUE
)
}
\arguments{
\item{data}{A tsibble object}

\item{y}{Column name of the variable of interest}

\item{stationary}{If \code{TRUE}, use \code{report_ndiffs()} to differentiate series}

\item{lags}{Numeric vector indicating how many lags should be used to expand the data.}

\item{fill_down_na}{If \code{TRUE}, fill NA observations with the last observed value downwards.}

\item{seasonal_dummies}{If \code{TRUE}, add seasonal dummies to the data.}
}
\value{
A tsibble
}
\description{
This function performs a number of conditional treatments (see arguments) commonly useful for use in time series forecasting models.
}
\examples{
df_macro <- get_data(quiet = TRUE)
df_macro_augmented <- ts_transform(df_macro, "ipca")
}
