% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{ym_label}
\alias{ym_label}
\title{Format x-axislabels (dates) as year/month in two lines}
\usage{
ym_label(x)
}
\arguments{
\item{x}{Date vector, usually takes as input the output of \code{breaks} in \code{ggplot2::scale_x_date}}
}
\value{
Character vetor
}
\description{
Format x-axislabels (dates) as year/month in two lines
}
\examples{
dates <- seq.Date(Sys.Date(), Sys.Date() + 120, by = "month")
ym_label(dates)
}
